chatespe/requirements.txt
faiss-cpu==1.7.2
numpy==1.23.1
pandas==1.4.3
rapidfuzz==2.1.2
sentence-transformers==2.2.2
flask==2.1.2
flask-cors==3.0.10
flask-restful==0.3.9
requests==2.28.1
numpy==1.23.1
faiss-cpu==1.7.2
transformers==4.20.1
torch==1.12.0
scikit-learn==1.1.1
pandas==1.4.3